==> 05_MoveTraits.V0.1_IllustrativeFigure.Rmd <==
---
title: "MoveTraits.V0.1 Manuscript Figures"
author: "Anne G. Hertel"
date: "2/26/2025"
output: 
  html_document:
    theme: cosmo
    toc: yes
    toc_float:
      collapsed: true
    code_folding: show
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r packages, warning=FALSE,message=FALSE}
library(tidyverse); library(mapview)
library(ggridges);library(sp); library(hrbrthemes);library(Hmisc);library(showtext); library(bit64)
```

# Exploration of MoveTraits database

First exploration of the MoveTraits database compiled from open access data from the Tucker et al. 2023 "Behavioral responses of terrestrial mammals to COVID-19 lockdowns" (https://zenodo.org/records/7704108, file "Tucker_Road_Spatial.rds") and from all openly available GPS datasets in Movebank (efefctive 06.2024). We only used data from 2019 (i.e., not from 2020 during COVID lockdowns). 

```{r}
pathTOfolder <- "./DATA/MoveTraitsData/8.MoveTraits_db/"

MoveTrait.v0.1 <- readRDS(paste0(pathTOfolder,"MoveTrait.v0.1_individual.sum_20251011.rds"))

MoveTrait.v0.1$study_id <- as.integer64(MoveTrait.v0.1$study_id)
```

## Quick summary
We collected movement traits from a total of 97 bird and 52 mammal species represented by 5873 individuals (3294 birds, 2579 mammals). 
Data of 4096 individuals were collected from Movebank and 1777 from Tucker et al.

```{r,echo=F}
MoveTrait.v0.1 |>  tally()
MoveTrait.v0.1 |> filter(!duplicated(species)) |> group_by(class) |>  tally()
MoveTrait.v0.1 |> group_by(class) |>  tally()
MoveTrait.v0.1 |> group_by(source) |>  tally()
```

## where are studys from (spatial extent)
```{r,echo=F}
mapMoveTrait.v0.1 <- MoveTrait.v0.1  |>  
  filter(!is.na(mean.longitude)) |>  
  filter(!is.na(mean.latitude)) 

mapMoveTrait.v0.1 <- sf::st_as_sf(mapMoveTrait.v0.1, coords = c("mean.longitude", "mean.latitude"))
sf::st_crs(mapMoveTrait.v0.1) <- 4326
```

Overview of the study locations included in the current database
```{r,echo=F}
mapviewOptions(fgb = FALSE)

mapview(mapMoveTrait.v0.1[mapMoveTrait.v0.1$class=="mammal",],zcol = "common_name",
layer.name = 'Mammals')

mapview(mapMoveTrait.v0.1[mapMoveTrait.v0.1$class=="bird",],zcol = "common_name",
layer.name = 'Birds')
```

Interesting - across all temporal scales, the mean displacement distances and mean range size IN MAMMALS are highly correlated, Intensity of Use and diurnality are not correlated with space use per-se. 

In birds the correlations with range size are much weaker - likely many migratory species. 

# Figure 1 - morphologic proxy trait versus displacement - MAMMALS

```{r,echo=F,warning=FALSE,message=FALSE}
pant <- read.table("./DATA/species_traits/PanTHERIA_1-0_WR05_Aug2008.txt",
                   header=T,sep = "\t")
```

## Fig1A - log body mass versus log monthly MCP

Get mule deer shape
```{r}
library(rphylopic)
uuid <- get_uuid(name = "Odocoileus hemionus")
img <- get_phylopic(uuid = uuid)
```

```{r}
MoveTrait.v0.1.sp <- readRDS(paste0(pathTOfolder,"MoveTrait.v0.1_species.sum_20251011.rds"))
```

```{r}
MoveTrait.v0.1.sp <- MoveTrait.v0.1.sp |> 
  mutate(common_name = Hmisc::capitalize(common_name))

MoveTrait.v0.1.sp <- MoveTrait.v0.1.sp |> 
  filter(movement.mode %nin% c("arboreal","swim")) |> 
  filter(!(class == "mammal" & movement.mode == "fly"))
```


```{r, fig.showtext = TRUE}
cols <- c("1" = "#228833", "2" = "#4477AA", "3" = "#AA3377")

 tot <- c("White-tailed deer","African elephant", "South american tapir","African forest elephant","Reindeer/caribou","Nyala","Raccoon","Blue wildebeest","Bobcat","Fisher","Cape bushbuck","Chamois","Kinkajou","Cougar","Golden jackal","Mule deer")
 
 below.l <- c("White-tailed deer","Red deer","Nyala","Raccoon","Blue wildebeest","Bobcat","Fisher","Cougar")
 left <- c("African elephant","African forest elephant", "Reindeer/caribou")
 right <- c("Cape bushbuck","Chamois","Kinkajou","South american tapir","Golden jackal")
 bold <- c("Mule deer")

dat.FigA <-
  MoveTrait.v0.1.sp |> 
  filter(class == "mammal") |> 
  dplyr::select(mcp1m.mean,common_name,species) |> 
  filter(!is.na(mcp1m.mean)) |>
  dplyr::mutate(label_above = ifelse(common_name %in% tot, "", common_name),
                  label_below = ifelse(common_name %in% below.l, common_name, ""),
                  label_left = ifelse(common_name %in% left, common_name, ""),
                  label_right = ifelse(common_name %in% right, common_name, ""),
                label_bold = ifelse(common_name %in% bold, common_name, "")) |>
  left_join(pant[,c("MSW05_Binomial","X6.2_TrophicLevel","X5.1_AdultBodyMass_g")], 
            by = c("species" = "MSW05_Binomial")) |> 
  mutate(X6.2_TrophicLevel = as.character(X6.2_TrophicLevel)) |>
  mutate(AdultBodyMass_kg = X5.1_AdultBodyMass_g/1000) |> 
  mutate(mcp1m.mean = mcp1m.mean* 0.000001)

# rename trophic level for missing species
  dat.FigA[dat.FigA$X6.2_TrophicLevel %in% "-999","species"]
  dat.FigA[dat.FigA$species %in% c("Equus quagga",
                                   "Lepus europaeus",
                                   "Capreolus capreolus"),"X6.2_TrophicLevel"] <- "1"
  dat.FigA[dat.FigA$species %in% c("Papio anubis"),"X6.2_TrophicLevel"] <- "2"

  dat.FigA[is.na(dat.FigA$X6.2_TrophicLevel),"species"]
  dat.FigA[dat.FigA$species %in% c("Tragelaphus sylvaticus"),"X6.2_TrophicLevel"] <- "1"
  
  dat.FigA[is.na(dat.FigA$AdultBodyMass_kg),"species"]
  dat.FigA[dat.FigA$species %in% c("Tragelaphus sylvaticus"),"AdultBodyMass_kg"] <- 50


  m1b <- lm(log(mcp1m.mean) ~ log(AdultBodyMass_kg),dat.FigA)
  summary(m1b)
  
FigA <-
  ggplot(dat.FigA, 
         aes(x = log(AdultBodyMass_kg), y = log(mcp1m.mean))) + 
  geom_smooth(method='lm', col = "black",se=F) +
  geom_point(aes(color = factor(X6.2_TrophicLevel))) +
  labs(x = "Mean body mass (kg)",
        y =  expression ("Monthly range size"~(km^2)),
        title = "Species level", tag = "A") +
  geom_text(aes(label = label_above),
            size = 1.8, position = position_dodge(width = 1),
            vjust = -0.55,hjust = -0.12) +
  geom_text(aes(label = label_below),
            size = 1.8,position = position_dodge(width = 1),
            vjust = 1.2,hjust = 1.1) +
  geom_text(aes(label = label_left),
            size = 1.8,position = position_dodge(width = 1),
            vjust = -0.55,hjust = 0.8) +
  geom_text(aes(label = label_right),
            size = 1.8,position = position_dodge(width = 1),
            vjust = 0.5,hjust = -0.12) +
  geom_text(aes(label = label_bold),
            size = 2.2,position = position_dodge(width = 1),
            vjust = 0.5,hjust = -0.12, fontface = "bold") +
  scale_x_continuous(limits = c(0.8,8.6),
                     breaks = c(log(5),log(50),log(500),log(5000)),
                     labels = c(5,50,500,5000)) +
  scale_y_continuous(breaks = c(log(0.1),log(1),log(10),log(100), log (1000)),
                     labels = c(0.1, 1,10,100, 1000)) +
  geom_text(x = 7.8, y = 2, size = 3.5,
            label = expression(R^2 ~ "= 0.20")) +
  scale_color_manual(values=cols, name="",labels = c("Herbivore", "Omnivore","Carnivore"),
                     na.translate = F) +
  theme_bw(base_size = 11, base_family = "Avenir") +
  theme(legend.position = c(0.8,0.22), #c(0.8, 0.2)
        legend.background = element_rect(fill="transparent"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.title=element_text(size=10),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        plot.title = element_text(face = "bold"),
        legend.spacing.x = unit(0.2, 'cm'),
        legend.spacing.y = unit(0.1, 'cm'),
        legend.text = element_text(margin = margin(l = -3)),
        plot.tag.position = c(0.02, 0.99),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points"),
        plot.tag = element_text(face = "bold") ) +     
  add_phylopic(img = img, x = 4.8, y = 3.7, height = 0.7)
```

Based on 41 species

# Figure 2 - inter-individual variability

Only including species with at least 10 individuals!! 27 species

## 24hr displacement distance
```{r,echo=F,warning=FALSE,message=FALSE,fig.showtext = TRUE}
MoveTrait.v0.1 <- MoveTrait.v0.1 |> 
  mutate(common_name = recode(common_name, "reindeer" = "reindeer/caribou")) |> 
  mutate(common_name = recode(common_name, "elk" = "red deer/elk")) |> 
  mutate(common_name = recode(common_name, "red deer" = "red deer/elk"))

dat.FigB <-
  MoveTrait.v0.1 |> 
  filter(class == "mammal") |> 
    left_join(pant, by = c("species" = "MSW05_Binomial")) |> 
  filter(d24h.mean < 30000) |> 
  dplyr::select(d24h.mean,common_name,species,X6.2_TrophicLevel) |> 
  filter(!is.na(d24h.mean)) |> 
  mutate(d24h.mean = d24h.mean/1000) |> 
  group_by(common_name) |> 
  filter(n() > 9) |> 
  mutate(species.medD24h=median(d24h.mean)) |> 
  ungroup() |> 
  droplevels()

# assign trophic level for missing species
dat.FigB[!duplicated(dat.FigB$species) & dat.FigB$X6.2_TrophicLevel %in% "-999","species"]
dat.FigB[!duplicated(dat.FigB$species) & is.na(dat.FigB$X6.2_TrophicLevel),"species"]

  dat.FigB[dat.FigB$species %in% c("Equus quagga",
                                   "Lepus europaeus"),"X6.2_TrophicLevel"] <- 1
  dat.FigB[dat.FigB$species %in% c("Papio anubis"),"X6.2_TrophicLevel"] <- 2

dat.FigB <- dat.FigB |> 
  mutate(common_name = Hmisc::capitalize(common_name))

library(ggtext)

dat.FigB <-
  dat.FigB |> 
  mutate(
    BoldLabel = str_starts(common_name, "Mu"),
    common_name_label = if_else(BoldLabel, glue::glue("**{common_name}**"), glue::glue("{common_name}")))

FigB <- 
  ggplot(dat.FigB, aes(x = d24h.mean, y = reorder(common_name_label, species.medD24h))) + 
  geom_density_ridges(aes(fill = factor(X6.2_TrophicLevel)),quantile_lines = T, 
                      quantiles = 2,  
                      color = NA, vline_color = "white", 
                      jittered_points = TRUE,
                      position = position_points_jitter(width = 0.05, height = 0),
                      point_shape = '|', point_size = 1.5, point_alpha = 1, alpha = 0.7,
                      point_color ="black", rel_min_height = 0.01, )+
  labs(x = "Individual mean daily displacement (km)",
        title = "Individual level",
        tag = "B") +
  geom_vline(xintercept=0) + 
  scale_fill_manual(values=cols,name="",labels = c("Herbivore", "Omnivore","Carnivore")) + 
  scale_x_continuous(limits =c(0, 17), expand=c(0,0))+  
  theme_bw(base_size = 11, base_family = "Avenir") +
    theme(legend.position = c(0.8, 0.2),
        legend.background = element_rect(fill="transparent", color = NA),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.box = "none",
        plot.title = element_text(face = "bold"),
        axis.title.y = element_blank(),
        panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points"),
        plot.tag.position = c(0.02, 0.99),
        plot.tag = element_text(face = "bold"),
        axis.text.y = ggtext::element_markdown()) +   
  add_phylopic(img = img, x = 14, y = 12, height = 1)
```


# Figure 3 - interindividual

```{r}
MoveTrait.v0.1.sp <- readRDS(paste0(pathTOfolder,"MoveTrait.v0.1_withinindividual_20251011.rds"))
```

```{r}
d24h.raw <- MoveTrait.v0.1_spatial |> 
  unnest(displ.24h) |> 
  dplyr::select("individual_id","species","common_name",
                "d24h","t_","x_","y_","contact_person_name","source") |> 
  rename(x = x_)|> 
  rename(y = y_)

mule <- d24h.raw[d24h.raw$common_name %in% "mule deer",]

tr <- mule |>
  group_by(individual_id) |>
  mutate(d24h = d24h/1000) |>
  filter(n() >= 55) |>
  mutate(ind.medD24h=median(d24h)) |>
  ungroup() |>
  filter(individual_id %in% c("MD18F0424_Odocoileus hemionus","MD18F0518_Odocoileus hemionus",
                              "MD16F0053_Odocoileus hemionus", "MD19F0059_Odocoileus hemionus",
                              "MD17F0220_Odocoileus hemionus"))

FigC <- 
ggplot(tr, aes(x = d24h , y = fct_reorder(individual_id, ind.medD24h))) + 
  geom_density_ridges(fill = "#228833" ,quantile_lines = T, 
                      quantiles = 2,  
                      color = NA, vline_color = "white", 
                      jittered_points = TRUE,
                      position = position_points_jitter(width = 0.05, height = 0),
                      point_shape = '|', point_size = 1.5, point_alpha = 1, alpha = 0.7,
                      point_color ="black", rel_min_height = 0.01, )+
  labs(x = "Daily displacement (km)",
        y = "Individual",
        title = "Within-individual level",
       tag = "C") +
    theme_bw(base_size = 11, base_family = "Avenir") +
    theme(legend.position = c(0.8, 0.15),
        legend.background = element_rect(fill="transparent", color = NA),
        plot.title = element_text(face = "bold"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.box = "none",axis.text.y = element_blank(),
        panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points"), plot.tag.position = c(0.02, 0.99),
        plot.tag = element_text(face = "bold"))+
  add_phylopic(img = img, x = 4.2, y = 2, height = 0.5)

```


# Figure 4 - reaction norm

This needs to be developed a lot more!!!! Any ideas welcome - I was thinking along the lines of showing movement over a temporal gradient, essentially depicting migratory patterns - but not so easy. Any other ideas? Could also try other movement metrics - e.g. max displacement over 7 days??

Would be cool to include a temperature gradient.

```{r}
MoveTrait.v0.1_spatial <- readRDS(paste0(pathTOfolder,"MoveTrait.v0.1_withinindividual_20251011.rds"))
```

```{r,echo=F}
d24h.raw <- MoveTrait.v0.1_spatial |> 
  unnest(displ.24h) |> 
  dplyr::select("individual_id","species","common_name",
                "d24h","t_","x_","y_","contact_person_name","source") |> 
  rename(x = x_)|> 
  rename(y = y_)
```

Mule deer 24hr displacement against HFI

```{r}
mule <- d24h.raw[d24h.raw$common_name %in% "mule deer",]
```

Extract human footprint downloaded from https://doi.org/10.7927/H46T0JQ4

```{r}
library(terra)
hfi<-rast("/Users/ahertel/Documents/Work/GIS/HumanFootprintIndex/V3/wildareas-v3-2009-human-footprint.tif")
d24h.mule.sf <- sf::st_as_sf(mule, coords = c(6:7))
sf::st_crs(d24h.mule.sf) <- 4326
vect_d24h <- vect(d24h.mule.sf)
reprojected_vect_d24h <- project(vect_d24h, crs(hfi))
extracted_values <- extract(hfi, reprojected_vect_d24h)

mule$hfi <- extracted_values[,2]
```

Select mule deer that experience a high HFI
```{r}
mule.high.hfi <- mule |> 
  group_by(individual_id) |> 
  mutate(hfi.max = max(hfi,na.rm=T)) |> 
  ungroup() |> 
  filter(hfi.max > 20) |> 
  droplevels()
```

We use pairwise comparisons of two mule deer at a time nto identify pairs for which individuals adjust movement (i.e.,  daily displacement distance) differentially to HFI.

```{r, eval = F}
# Get all unique pairwise comparisons of factor levels
factor_levels <- unique(mule.high.hfi$individual_id)
pairwise_comparisons <- combn(factor_levels, 2, simplify = FALSE)

# Initialize an empty dataframe to store results
model.results <- data.frame(
  Comparison = character(),
  P_int = numeric(),
  P_ID = numeric(),
  stringsAsFactors = FALSE
)

# Loop through each pairwise comparison
for (pair in pairwise_comparisons) {
  
  # Subset data for the current pair of factor levels
  subset_data <- mule.high.hfi %>% filter(individual_id %in% pair)
  
  # Fit the linear model
  model <- lm(d24h ~ hfi * individual_id, data = subset_data)
  
  # Extract the p-value for the factor variable
  p_value_int <- round(summary(model)$coefficients[3, 4],3)
  p_value_ID <- round(summary(model)$coefficients[4, 4],3)
  
  # Save the results
  model.results <- rbind(model.results, data.frame(
    Comparison = paste(pair, collapse = " vs "),
    ID1 = pair[1],
    ID2 = pair[2],
    P_int = p_value_int,
    P_ID = p_value_ID,
    stringsAsFactors = FALSE
  ))
}

# View the results dataframe
p.sign <- model.results |> 
  filter(P_ID < 0.05 & P_int < 0.05)
```

For 258 out of 1653 pairwise comparisons, the two individuals in question move significantly different along the HFI gradient. 

```{r, eval = F}
significant_comparisons <- p.sign$Comparison

# Subset the pairwise_comparisons list using the significant comparisons
filtered_pairwise_comparisons <- pairwise_comparisons[
  sapply(pairwise_comparisons, function(pair) paste(pair, collapse = " vs ") %in% significant_comparisons)
]

# Create a new dataframe to store pairwise comparison data
comparison_data <- data.frame()

for (pair in filtered_pairwise_comparisons) {
  # Subset data for the current pair of factor levels
  subset_data <- mule.high.hfi %>% filter(individual_id %in% pair)
  
  # Add a column indicating the comparison
  subset_data$Comparison <- paste(pair, collapse = " vs ")
  
  # Combine the subset data into the comparison_data dataframe
  comparison_data <- rbind(comparison_data, subset_data)
}

# Plot significant comparisons (exemplary last 20 sign comparisons)
ggplot(comparison_data[comparison_data$Comparison %in% levels(factor(comparison_data$Comparison))[248:270],],
       aes(x = hfi, y = d24h, group = individual_id,color = individual_id)) + 
  geom_point()+
  geom_smooth(method = "lm")+
  facet_wrap(~Comparison)+
  theme(legend.position = "none") + xlim(0,50) +
  ylim(0,10000) 
```

Select two mule deer with different reactions to HFI as illustrative example 
```{r}
mule.high.hfi <- mule |> 
  filter(individual_id %in% c("706099A_Odocoileus hemionus", "MD19F0096_Odocoileus hemionus")) |> 
  droplevels()
```

```{r}
mapviewOptions(fgb = FALSE)

map.mule <- sf::st_as_sf(mule.high.hfi, coords = c("x", "y"))
sf::st_crs(map.mule) <- 4326

mapview(map.mule,
        zcol = "individual_id")
```


```{r}
model1 <- lm(d24h ~ hfi * individual_id, data = mule.high.hfi)
model2 <- lm(d24h ~ hfi + individual_id, data = mule.high.hfi)

anova(model2, model1, test = "LRT")

summary(model1)

prd <- expand.grid(hfi = seq(min(mule.high.hfi$hfi),max(mule.high.hfi$hfi),5),
            individual_id = levels(factor(mule.high.hfi$individual_id)))
prd$d24h <- predict(model1,prd)

```

Plot Reaction norm Panel CD
```{r,fig.showtext = TRUE}
FigD <-
ggplot(data      = mule.high.hfi,
       aes(x     = hfi,
           y     = d24h/1000,
           col   = individual_id,
           group = individual_id))+ #to add the colours for different classes
  geom_point(size     = 1.2,
             position = "jitter")+ #to add some random noise for plotting purposes
  theme_bw()+
  scale_color_manual(values = c("#332288","#44AA99"),
                     guide = guide_legend(title = NULL),
                         labels = c("Individual 1","Individual 2"))+
  geom_smooth(method = lm,
              se     = FALSE,
              size   = .5) +
  scale_x_continuous(limits = c(4,30),expand=c(0.01,0.01))+
  #scale_y_continuous(expand=c(0.01,0.01))+
  labs(x = "Human Footprint Index",
       y = "Daily displacement (km)",
       title = "Behavioral reaction norm",
        tag = "D") +
  theme_bw(base_size = 11, base_family = "Avenir") +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  theme(legend.position = c(.75,.85),
        legend.background = element_blank(),
        plot.title = element_text(face = "bold"),
        #axis.title.y = element_text(vjust = -2),
        plot.tag.position = c(0.02, 0.99),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points"),
        # ,plot.title = element_text(family = "Aptos-Display-Bold"),
        #  axis.title.x = element_text(family = "Aptos-Display-Bold"),
        #  axis.title.y = element_text(family = "Aptos-Display-Bold"),
        plot.tag = element_text(face = "bold")
        ) +
    add_phylopic(img = img, x = 16, y = 4, height = 0.5)


```

## Final Figure
```{r}
library(patchwork)

(free(FigA, type = "label")  / (FigB | (FigC / FigD))) +
  plot_layout(heights = unit(c(1, 2), c('null')))


ggsave("OUTPUT/Fig_ProofConcept.png", width = 19, height = 25, units = "cm",dpi = 1200)
```


# Supplemement - morphologic proxy trait versus displacement - BIRDS

## birds - species average 24hr displacement distance versus kips index

```{r}
library(traitdata)
data(avonet)

kips <- avonet |>
  group_by(scientificNameStd) |> 
  summarise(Hand.wing.Index = mean(Hand.wing.Index,na.rm=T))

avonet.bt <- read.table("./DATA/species_traits/AVONET_BirdTree.txt",header=T,sep = "\t")
avonet.eb <- read.table("./DATA/species_traits/AVONET_eBird.txt",header=T,sep = "\t")
colnames(avonet.eb)[1]<-"Species3"

avonet.trophic <- rbind(avonet.bt[,c("Species3","Trophic.Level")],avonet.eb[,c("Species3","Trophic.Level")])
avonet.trophic <- avonet.trophic[!duplicated(avonet.trophic$Species3),]
```

72 species

```{r,echo=F,warning=FALSE,message=FALSE,fig.showtext = TRUE}
cols2 <- c("Herbivore" = "#228833", "Scavenger" = "#CCBB44", "Carnivore" = "#AA3377")

# need to filter out extreme outliers

dat.FigSA <-
  MoveTrait.v0.1.sp |> 
  filter(class == "bird") |> 
  dplyr::select(d1h.mean,common_name,species) |> 
  filter(!is.na(d1h.mean)) |> 
  filter(d1h.mean < 110000) |> 
  dplyr::mutate(outlier = ifelse(d1h.mean > 9800, common_name, NA)) |> 
  droplevels() |>  
  left_join(kips, by = c("species" = "scientificNameStd")) |> 
  left_join(avonet.trophic, by = c("species" = "Species3")) |> 
  filter(!is.na(Trophic.Level)) 

  m2 <- lm((d1h.mean/1000) ~ Hand.wing.Index,dat.FigSA)
  summary(m2)

FigSA <-
  ggplot(dat.FigSA, aes(x = Hand.wing.Index, y = d1h.mean/1000)) + 
  geom_point(aes(color = Trophic.Level)) + 
  geom_smooth(method='lm', col = "black",se=F) +
  geom_text(aes(label = outlier),
            size = 1.8, position = position_dodge(width = 1),
            vjust = -1,hjust = 0.5) +
  labs(x="Hand-wing Index",y="Mean hourly displacement (km)",tag="A", title = "Species level")+
    geom_text(x = 30, y = 7, size = 3.5,
            label = expression(R^2 ~ "= 0.08")) +
  theme_bw(base_size = 11, base_family = "Avenir") +
  theme(legend.position = "bottom",
        legend.title=element_blank()) +
  scale_color_manual(values=cols2, 
                     name="", 
                     breaks = c("Herbivore", "Scavenger","Carnivore"))+
  theme(legend.position = c(0.2, 0.9),
        legend.background = element_rect(fill="transparent"),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.title=element_text(size=10)) +
  scale_x_continuous(limits = c(22.5, 70),expand = c(0,0)) +
  scale_y_continuous(limits = c(0, 22),expand = c(0.01,0.01)) +
  guides(color = guide_legend(byrow = TRUE)) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        legend.spacing.y = unit(5, "cm"),
        legend.text = element_text(margin = margin(l = -3)),
        plot.tag.position = c(0.02, 0.99),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points")) 
```

## 24hr displacement distance

46 species

```{r,echo=F,warning=FALSE,message=FALSE,fig.showtext = TRUE}
FigSB <- 
  MoveTrait.v0.1 |> 
  filter(class == "bird") |> 
  filter(d24h.mean < 400000) |> 
  dplyr::select(d24h.mean,common_name,species) |> 
  filter(!is.na(d24h.mean)) |> 
  mutate(d24h.mean = d24h.mean/1000) |> 
  group_by(common_name,species) |> 
  filter(n() > 9) |> # at least 10 ids
  mutate(species.medD24h=median(d24h.mean)) |> 
  ungroup() |> 
  droplevels() %>%
  left_join(kips, by = c("species" = "scientificNameStd")) |> 
  left_join(avonet.trophic, by = c("species" = "Species3")) |> 
  mutate(common_name = Hmisc::capitalize(common_name)) %>%
  ggplot(., aes(x = d24h.mean, y = reorder(common_name, species.medD24h))) + 
  geom_density_ridges(aes(fill = Trophic.Level), 
                      quantile_lines = T, quantiles = 2,  
                      color = NA, vline_color = "white", 
                      jittered_points = TRUE,
                      position = position_points_jitter(width = 0.05, height = 0),
                      point_shape = '|', point_size = 1.5, point_alpha = 1, alpha = 0.7,
                      point_color ="black", rel_min_height = 0.01) +
  labs(x = "Individual mean daily displacement (km)",
        title = "Individual level",
       tag="B") +
  theme_minimal() +
  geom_vline(xintercept=0) + 
  scale_x_continuous(limits =c(0, 300), expand=c(0,0))+
  scale_fill_manual(values=cols2, name="", breaks = c("Herbivore", "Scavenger","Carnivore"))+
  theme_bw(base_size = 11, base_family = "Avenir") +
    theme(legend.position = c(0.7, 0.2),
        legend.background = element_rect(fill="transparent", color = NA),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.box = "none",
        axis.text=element_text(size=8),
        axis.title.y = element_blank())+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        plot.tag.position = c(0.02, 0.99),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points")
        )

```

## Monthly Intensity of use of raptors

```{r,echo=F}
MoveTrait.v0.1_spatial <- readRDS(paste0(pathTOfolder,"MoveTrait.v0.1_withinindividual_20251011.rds"))

IoU.1m.raw <- MoveTrait.v0.1_spatial |>
  unnest(iou.1m)|> 
  dplyr::select("study_id","individual_id","species","common_name","class","year","month","iou1m","contact_person_name") |> 
  filter(!is.na(iou1m))
```

## Monthly Intensity of use in migratory raptors
```{r,echo=F}
rap <- IoU.1m.raw |> 
  filter(class == "bird") |> 
  filter(iou1m < 200) |> 
  filter(common_name %in% c("eurasian griffon vulture")) 

rap$id_year <- paste(rap$individual_id,rap$year,sep="_")
```

```{r,fig.showtext = TRUE}
FigSC <-
  ggplot()+ 
  geom_point(data = rap[rap$id_year %in% 
                                c("186182027_2017" , "186182027_2023"),],
             aes(x = month,y = iou1m, col = id_year, group = id_year),
             size = 1.5, shape = 19)+
  geom_smooth(data = rap[rap$id_year %in%
                                c("186182027_2017" , "186182027_2023"),], 
              aes(x = month,y = iou1m, col = id_year, group = id_year),
              size   = .8, se = F) +
  scale_color_manual(values = c("#44AA99","#332288"),
                     guide = guide_legend(title = NULL),
                     labels = c("Subadult","Breeding"))+
  scale_x_continuous(expand=c(0.01,0.01))+
  labs(x = "Month",
       y = "Monthly intensity of use",
       title = "Within-individual level",
        tag = "C") +
  theme_bw(base_size = 11, base_family = "Avenir") +
  theme(legend.position = c(0.83,0.91))+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        legend.background = element_rect(fill = NA),
        plot.tag.position = c(0.02, 0.99),
        plot.margin = unit(c(5.5,5.5,5.5,1), "points"))
```

# Final SupplementaryFigure
```{r}
library(patchwork)

((FigSA / FigSC ) | FigSB) 
ggsave("OUTPUT/Fig_ProofConcept_Supplement.png", width = 19, height = 19, units = "cm")

```
